pyproject.toml
[tool.poetry]
name = "personal-assistant"
version = "0.1.0"
description = ""
authors = ["Logan Hillyer <shivastranded@proton.me>"]
readme = "README.md"
packages = [{include = "personal_assistant"}]

[tool.poetry.scripts]
main = "./main.py"

[tool.poetry.dependencies]
python = "3.9.16"
elevenlabs = "^0.2.7"
openai = "^0.27.5"
pydub = "^0.25.1"
pyglet = "^2.0.5"
python-dotenv = "^1.0.0"
toml = "^0.10.2"
SpeechRecognition = "^3.10.0"
PyAudio = "^0.2.13"
openai-whisper = "^20230314"
spacy = "^3.5.2"
langchain = "^0.0.172"
google-search-results = "^2.4.2"
wikipedia = "^1.4.0"
streamlit = "^1.22.0"
faiss-gpu = "^1.7.2"
rich = "^13.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
